local-backend/requirements.txt
python-pptx
pillow
boto3